Remove `ScrapePartnersCommand_V2`, update dependencies, and refactor `Makefile` and `console` setup.
