+++ composer.json
@@ -35,32 +35,32 @@
   },
   "require": {
     "php": "8.4.* || 8.5.*",
-    "ext-json": "*",
-    "ext-curl": "*",
     "ext-bcmath": "*",
+    "ext-curl": "*",
     "ext-dom": "*",
     "ext-intl": "*",
-    "psr/log": "^3",
-    "fig/http-message-util": "^1",
-    "giggsey/libphonenumber-for-php": "^8",
-    "darsyn/ip": "^5",
-    "nesbot/carbon": "^3",
-    "moneyphp/money": "^4",
+    "ext-json": "*",
     "bitrix24/b24phpsdk": "dev-v3-dev",
-    "doctrine/orm": "^3",
+    "darsyn/ip": "^5",
     "doctrine/doctrine-bundle": "*",
     "doctrine/doctrine-migrations-bundle": "*",
+    "doctrine/orm": "^3",
+    "fig/http-message-util": "^1",
+    "giggsey/libphonenumber-for-php": "^8",
     "knplabs/knp-paginator-bundle": "^6",
     "league/csv": "^9.0",
+    "moneyphp/money": "^4",
+    "nesbot/carbon": "^3",
     "odolbeau/phone-number-bundle": "^3.0",
+    "psr/log": "^3",
+    "symfony/cache": "^7",
+    "symfony/console": "^7",
+    "symfony/dotenv": "^7",
     "symfony/event-dispatcher": "^7",
     "symfony/http-client": "^7",
     "symfony/serializer": "^7",
     "symfony/uid": "^7",
-    "symfony/yaml": "^7",
-    "symfony/cache": "^7",
-    "symfony/console": "^7",
-    "symfony/dotenv": "^7"
+    "symfony/yaml": "^7"
   },
   "require-dev": {
     "lendable/composer-license-checker": "^1.2",
